client/controller: Move client specific stuff to index.web.js
